Add Header component and integrate it into App; refactor Layout to remove Get Started button logic
