+++ src/App.tsx
@@ -1,6 +1,7 @@
 import { BrowserRouter as Router, Routes, Route, useLocation } from 'react-router-dom'
 import { AnimatePresence } from 'framer-motion'
 import BackgroundWrapper from './Components/ui/background'
+import Header from './Components/ui/Header'
 import PageTransition from './Components/ui/PageTransition'
 import './App.css'
 import Landing from './Components/pages/landing'
@@ -54,6 +55,7 @@ function App() {
   return (
     <Router>
       <BackgroundWrapper>
+        <Header />
         <AnimatedRoutes />
       </BackgroundWrapper>
     </Router>
